Deleted zoom buttons from map and button plus. True Magic Numbers.
