+++ WeatherProjectTestOrion/Constants/CitiesStorage.swift
@@ -7,37 +7,37 @@
 
 import Foundation
 struct CitiesStorage {
-  static let citiesStorage = [
-    "Moscow",
-    "Saint Petersburg",
-    "Novosibirsk",
-    "Yekaterinburg",
-    "Kazan",
-    "Nizhny Novgorod",
-    "Chelyabinsk",
-    "Samara",
-    "Omsk",
-    "Rostov-on-Don",
-    "Ufa",
-    "Krasnoyarsk",
-    "Voronezh",
-    "Perm",
-    "Volgograd",
-    "Krasnodar",
-    "Saratov",
-    "Tyumen",
-    "Barnaul",
-    "Irkutsk",
-    "Sochi",
-    "Novokuznetsk"
-  ]
+    static let citiesStorage = [
+        "Moscow",
+        "Saint Petersburg",
+        "Novosibirsk",
+        "Yekaterinburg",
+        "Kazan",
+        "Nizhny Novgorod",
+        "Chelyabinsk",
+        "Samara",
+        "Omsk",
+        "Rostov-on-Don",
+        "Ufa",
+        "Krasnoyarsk",
+        "Voronezh",
+        "Perm",
+        "Volgograd",
+        "Krasnodar",
+        "Saratov",
+        "Tyumen",
+        "Barnaul",
+        "Irkutsk",
+        "Sochi",
+        "Novokuznetsk"
+    ]
 }
 struct PopularCitiesStorage {
-  static let storage = [
-    "Moscow",
-    "Kaliningrad",
-    "Saint Petersburg",
-    "Yekaterinburg",
-    "Omsk"
-  ]
+    static let storage = [
+        "Moscow",
+        "Kaliningrad",
+        "Saint Petersburg",
+        "Yekaterinburg",
+        "Omsk"
+    ]
 }
