+++ WeatherProjectTestOrion/Constants/ConstantsGrid.swift
@@ -1,0 +1,16 @@
+//
+//  ConstantsItems.swift
+//  WeatherProjectTestOrion
+//
+//  Created by Tatyana Ilvutikova on 12.01.2022.
+//
+
+import Foundation
+import SwiftUI
+
+enum Constants {
+    enum WidthGrid {
+        public static let maxTemperature: CGFloat = 55
+        public static let  minTemperature: CGFloat = 40
+    }
+}
