+++ WeatherProjectTestOrion/Constants/Images.swift
@@ -1,0 +1,13 @@
+//
+//  Images.swift
+//  WeatherProjectTestOrion
+//
+//  Created by Tatyana Ilvutikova on 21.01.2022.
+//
+
+import Foundation
+
+enum Images {
+    public static let backgroundHeaderWeatherView = "cloudyBackground"
+    public static let imageMapPin = "mappin"
+}
