+++ WeatherProjectTestOrion/Constants/MagicNumber.swift
@@ -1,0 +1,21 @@
+//
+//  CGFloat.swift
+//  WeatherProjectTestOrion
+//
+//  Created by Tatiana Ilvutikova on 21.01.2022.
+//
+
+import Foundation
+import SwiftUI
+
+struct MagicNumber {
+    static let x05: CGFloat = 2
+    static let x1: CGFloat = 4
+    static let x2: CGFloat = 8
+    static let x3: CGFloat = 12
+    static let x4: CGFloat = 16
+    static let x5: CGFloat = 20
+    static let x6: CGFloat = 24
+    static let x8: CGFloat = 32
+    static let x16: CGFloat = 64
+}
