+++ WeatherProjectTestOrion/Extensions+/Colors.swift
@@ -1,0 +1,25 @@
+//
+//  Colors.swift
+//  WeatherProjectTestOrion
+//
+//  Created by Tatyana Ilvutikova on 13.01.2022.
+//
+
+import SwiftUI
+
+extension Color {
+    static let mainText = Color("mainText")
+    static let mainTextWhite = Color("mainTextWhite")
+    static let subText = Color("subText")
+    static let measure = Color("measure")
+    static let borderCityRow = Color("borderCityRow")
+    static let iconWeatherCityRow = Color("iconWeatherCityRow")
+    static let showMap = Color("showMap")
+    static let popularCityBackground = Color("popularCitiesBackground")
+    static let popularCitiesOverlay = Color("popularCitiesOverlay")
+    static let buttonDeleteTextField = Color("buttonDeleteTextField")
+    static let buttonSheetCancel = Color("buttonSheetCancel")
+    static let searchBackground = Color("searchBackground")
+    static let sheetDescriptionWeather = Color("sheetDescriptionWeather")
+    static let currentLocation = Color("currentLocation")
+}
