--- WeatherProjectTestOrion/Extensions+Constants/CGFloat.swift
@@ -1,51 +1,0 @@
-//
-//  CGFloat.swift
-//  WeatherProjectTestOrion
-//
-//  Created by Tatiana Ilvutikova on 21.01.2022.
-//
-
-import Foundation
-import SwiftUI
-
-extension CGFloat {
-  static var main: CGFloat {
-    return 8
-  }
-  static var x2: CGFloat {
-    return 16
-  }
-  static var x3: CGFloat {
-    return 24
-  }
-  static var x4: CGFloat {
-    return 32
-  }
-  static var x8: CGFloat {
-    return 64
-  }
-  static var second: CGFloat {
-    return 2
-  }
-  static var third: CGFloat {
-    return 3
-  }
-  static var s2: CGFloat {
-    return 4
-  }
-  static var s3: CGFloat {
-    return 6
-  }
-  static var s4: CGFloat {
-    return 8
-  }
-  static var s5: CGFloat {
-    return 10
-  }
-  static var s6: CGFloat {
-    return 12
-  }
-  static var s11: CGFloat {
-    return 22
-  }
-}
