+++ WeatherProjectTestOrion/Main/View/MainView.swift
@@ -7,25 +7,22 @@
 
 import SwiftUI
 struct MainView: View {
-  @State var isScrolled = false
-  @EnvironmentObject var weatherVM: WeatherViewModel
-  @EnvironmentObject var citiesVM: CitiesListViewModel
-  var body: some View {
-    VStack(spacing: .zero) {
-      HeaderWeatherView(
-        isScrolled: $isScrolled)
-        .frame(height: isScrolled ? 180 : 280)
-        .animation(.easeOut(duration: 0.3), value: isScrolled)
-      DailyWeatherListView(
-        isScrolled: $isScrolled)
+    @State var isScrolled = false
+    @EnvironmentObject var weatherVM: WeatherViewModel
+    @EnvironmentObject var citiesVM: CitiesListViewModel
+    var body: some View {
+        VStack(spacing: .zero) {
+            HeaderWeatherView(
+                isScrolled: $isScrolled)
+                .frame(height: isScrolled ? 180 : 280)
+                .animation(.easeOut(duration: 0.3), value: isScrolled)
+            DailyWeatherListView(
+                isScrolled: $isScrolled)
+        }
+        .onAppear {
+            Task {
+                await weatherVM.getAllWeather()
+            }
+        }
     }
-    .onAppear {
-      print("текущий город на главном экране", weatherVM.currentCity)
-      print("список городов", citiesVM.citiesList)
-      Task {
-        // MARK: - "если регион стоит как en_RU, то какой запрос делать?"
-        await weatherVM.getAllWeather()
-      }
-    }
-  }
 }
